AT-109: validate go-base is up to date and ready to use for new services (#4)
Polishing up the documentation. Calling out the format of the repository name to use and ensuring consistency in the initial setup. Use `master` as the origin branch name for now. It appears cloning from the `hello-go` template uses `master` instead of `main`. At some point, GitHub will change this behavior; we can update this doc accordingly then. Finally, put a placeholder in the spot for the new service description.